+++ .gitignore
@@ -1,2 +1,2 @@
-bin/Debug/net7.0/MapCycle/
-bin/Debug/net7.0/MapCycle.zip
+bin/
+obj/
