--- MapCycle.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MapCycle", "MapCycle.csproj", "{96BCA30B-7491-431A-9B18-DA1F00B150E1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{96BCA30B-7491-431A-9B18-DA1F00B150E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{96BCA30B-7491-431A-9B18-DA1F00B150E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{96BCA30B-7491-431A-9B18-DA1F00B150E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{96BCA30B-7491-431A-9B18-DA1F00B150E1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {04FF030B-EC7B-47B6-94E3-D52DB95B3401}
-	EndGlobalSection
-EndGlobal
